+++ tests/testthat/test-jitter_power.R
@@ -11,8 +11,7 @@ test_that("jitter_power throws error with invalid arguments", {
                  sim_total = 2,
                  samp_control = "MVN",
                  s_control = .01, 
-                 upper_tail = 0.995,
-                 lower_tail = 0.005, 
+                 alpha = 0.01, 
                  resolution = 10, 
                  edge = "uniform")
   )
@@ -22,8 +21,7 @@ test_that("jitter_power throws error with invalid arguments", {
                  sim_total = NULL,
                  samp_control = "MVN",
                  s_control = .01, 
-                 upper_tail = 0.995,
-                 lower_tail = 0.005, 
+                 alpha = 0.01,
                  resolution = 10, 
                  edge = "uniform")
   )
@@ -33,8 +31,7 @@ test_that("jitter_power throws error with invalid arguments", {
                  sim_total = 2,
                  samp_control =  NULL,
                  s_control = .01, 
-                 upper_tail = 0.995,
-                 lower_tail = 0.005, 
+                 alpha = 0.01,
                  resolution = 10, 
                  edge = "uniform")
   )
@@ -44,8 +41,7 @@ test_that("jitter_power throws error with invalid arguments", {
                  sim_total = 2,
                  samp_control = "MVN",
                  s_control = NULL, 
-                 upper_tail = 0.995,
-                 lower_tail = 0.005, 
+                 alpha = 0.01, 
                  resolution = 10, 
                  edge = "uniform")
   )
@@ -57,9 +53,8 @@ test_that("jitter_power works", {
   expect_named(jitter_power(obs_data = unique(chorley),
                             sim_total = 2,
                             samp_control = "MVN",
-                            s_control = .01, 
-                            upper_tail = 0.995,
-                            lower_tail = 0.005, 
+                            s_control = 0.01, 
+                            alpha = 0.01, 
                             resolution = 10, 
                             edge = "uniform")
   )